requirements.txt
httpx==0.27.2
websockets==13.1
pydantic==2.9.2
python-dotenv==1.0.1
pyyaml==6.0.2
uvloop==0.20.0; sys_platform == 'linux'
rich==13.9.4
structlog==24.4.0
